frontend: session info: applied Yu PR review
